requirements.txt
fastapi
uvicorn
python-multipart
python-dotenv
rapidocr-onnxruntime
pillow
